Working wp-gulp
This is a Wordpress modification to the gulp file.js which is explained at https://coursetro.com/posts/code/130/Learn-Bootstrap-4-Final-in-2018-with-our-Free-Crash-Course

You will be required to add the external dependancies from bootstrap and your lease be required to add the npm dependences, all of which are described and explains in the above link.

Copy the files to themes directory. Within your Wordpress theme directory and create a sub directory /src in the theme you create. Within the /src dir add three more directories, /css, /js and /scss


wp-content
|/themes
|-/gulpfile.js
|-/twentyseventeen
|-/YourTheme
|-|-/index.php
|-|-/src
|-|-|-/css
|-|-|-/js
|-|-|-/scss
|-|-|-|-/styles.scss